tasks/requirements.txt
linearmodels==6.0
matplotlib==3.8.4
networkx==3.3
numpy==2.1.1
pandas==2.2.3
powerlaw==1.5
pyarrow==17.0.0
pyfixest==0.24.2
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1
statsmodels==0.14.2
